cleanup

+++ enet-csharp/ENet/plugins/NativeSockets/Unix/Shared/c/iOSSocketPal.cs
@@ -74,9 +74,9 @@ namespace NativeSockets
         public static extern int getnameinfo(sockaddr* pSockaddr, int SockaddrLength, ref byte pNodeBuffer, ulong NodeBufferSize, byte* pServiceBuffer, ulong ServiceBufferSize, int Flags);
 
         [DllImport(NATIVE_LIBRARY, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int sendmsg(int sockfd, msghdr* msg, int flags);
+        public static extern nint sendmsg(int sockfd, msghdr* msg, int flags);
 
         [DllImport(NATIVE_LIBRARY, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int recvmsg(int sockfd, msghdr* msg, int flags);
+        public static extern nint recvmsg(int sockfd, msghdr* msg, int flags);
     }
 }
